Fixed 32-bit/64-bit unsigned long mismatch between Windows and Linux on LSD Radix Sort - switched to just unsigned which is the same size on Windows and Linux. Got Linux to build and run with g++. Got Windows to build with VS2022 and Intel compilers.

+++ RadixSortCommon.h
@@ -49,9 +49,9 @@ inline unsigned __int64 logicalRightShift_ru( __int64 a, unsigned long shiftAmou
 }
 #endif
 template< class _Type >
-inline unsigned long extractDigit( _Type a, _Type bitMask, unsigned long shiftRightAmount )
+inline unsigned extractDigit( _Type a, _Type bitMask, unsigned shiftRightAmount )
 {
-	unsigned long digit = (unsigned long)(( a & bitMask ) >> shiftRightAmount );	// extract the digit we are sorting based on
+	unsigned digit = (unsigned)(( a & bitMask ) >> shiftRightAmount );	// extract the digit we are sorting based on
 	return digit;
 }
 template< class _Type >
